added fun to disconnect and connect w/out restarting added labels to attributes in ui changed ui layout in last sync and dir sections and etc

+++ app/src/main/java/com/example/syncapp/ui/SyncUiState.kt
@@ -15,7 +15,7 @@ data class SyncUiState(
     var localFilesCount: Int = 0,
     var globalFilesCount: Int = 0,
     var globalDirsCount: Int = 0,
-    var globalSize: Int = 0,
+    var globalSize: Long = 0,
     var localDirsCount: Int = 0,
     var localSize: Int = 0,
     var isUpToDate: Boolean = false,
